+++ src/main/kotlin/com/github/intheclouddan/intellijpluginld/FlagStore.kt
@@ -11,6 +11,7 @@ import com.github.intheclouddan.intellijpluginld.notifications.ConfigNotifier
 import com.github.intheclouddan.intellijpluginld.settings.LaunchDarklyMergedSettings
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import com.intellij.openapi.application.invokeLater
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
@@ -134,11 +135,14 @@ class FlagStore(private var project: Project) {
             try {
                 val ldProject = LaunchDarklyApiClient.projectInstance(project, settings.authorization).getProject(settings.project)
                 val myStreamBaseURI = settings.baseUri.replace("app", "stream")
-                val (store, client) = createClientAndGetStore(ldProject.environments.find { it.key == settings.environment }!!.apiKey, myStreamBaseURI)
-                flagStore = store!!
-                flagClient = client
-                flagTargeting(store)
-                flagListener(client, store)
+                invokeLater {
+                    val (store, client) = createClientAndGetStore(ldProject.environments.find { it.key == settings.environment }!!.apiKey, myStreamBaseURI)
+                    flagStore = store!!
+                    flagClient = client
+                    flagTargeting(store)
+                    flagListener(client, store)
+                }
+
             } catch (err: ApiException) {
                 val notify = ConfigNotifier()
                 notify.notify(project, "Project: ${settings.project} Error: $err")