+++ src/main/kotlin/com/github/intheclouddan/intellijpluginld/toolwindow/FlagNodeParent.kt
@@ -2,6 +2,7 @@ package com.github.intheclouddan.intellijpluginld.toolwindow
 
 import com.github.intheclouddan.intellijpluginld.FlagStore
 import com.github.intheclouddan.intellijpluginld.LDIcons
+import com.github.intheclouddan.intellijpluginld.featurestore.FlagConfiguration
 import com.intellij.ide.projectView.PresentationData
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
@@ -14,7 +15,8 @@ class FlagNodeParent(FFlag: FeatureFlag, private var flags: FeatureFlags, myProj
     private var children: MutableList<SimpleNode> = ArrayList()
     private val getFlags = myProject.service<FlagStore>()
     var flag: FeatureFlag = FFlag
-    var env = getFlags.flagConfigs[flag.key]!!
+    var env = getFlags.flagConfigs[flag.key]
+            ?: FlagConfiguration(flag.key, null, null, listOf(), listOf(), arrayOf(), false, -1)
     val key: String = flag.key
 
 
@@ -57,9 +59,11 @@ class FlagNodeParent(FFlag: FeatureFlag, private var flags: FeatureFlags, myProj
 
     override fun update(data: PresentationData) {
         super.update(data)
-        env = getFlags.flagConfigs[flag.key]!!
+        env = getFlags.flagConfigs[flag.key]
+                ?: FlagConfiguration(flag.key, null, null, listOf(), listOf(), arrayOf(), false, -1)
         flag = getFlags.flags.items.find { it.key == flag.key }!!
-        val enabledIcon = if (env.on) LDIcons.TOGGLE_ON else LDIcons.TOGGLE_OFF
+        // Flag version should only be -1 if we manually created a FlagConfiguration so set icon to warning.
+        val enabledIcon = if (env.version === -1) LDIcons.TOGGLE_DISCONNECTED else if (env.on) LDIcons.TOGGLE_ON else LDIcons.TOGGLE_OFF
         val label = flag.name ?: flag.key
         data.presentableText = label
         data.setIcon(enabledIcon)